pyproject.toml
[project]
name = "music-tag-app"
version = "0.1.0"
description = "本地音乐 Metadata 管理 Web App"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "mutagen>=1.47",
    "httpx>=0.27",
    "Pillow>=10.0",
    "python-multipart>=0.0.9",
    "pyncm>=1.0",
]
